dtyping

+++ python/src/spark_rapids_ml/umap.py
@@ -509,7 +509,7 @@ class UMAPModel(_CumlModel, UMAPClass, _UMAPCumlParams):
     ) -> Tuple[_ConstructFunc, _TransformFunc, Optional[_EvaluateFunc],]:
         cuml_alg_params = self.cuml_params.copy()
         driver_embedding = self.embedding_
-        driver_raw_data = np.array(self.raw_data_)
+        driver_raw_data = np.array(self.raw_data_, dtype=np.float32)
 
         def _construct_umap() -> CumlT:
             import cupy as cp
@@ -530,9 +530,9 @@ class UMAPModel(_CumlModel, UMAPClass, _UMAPCumlParams):
                 )
 
             internal_model = CumlUMAP(**cuml_alg_params)
-            internal_model.embedding_ = (
-                cp.array(driver_embedding).astype(cp.float32).data
-            )
+            internal_model.embedding_ = cp.array(
+                driver_embedding, dtype=cp.float32
+            ).data
             internal_model._raw_data = raw_data_cuml
 
             return internal_model
